--- plugins/data-management/src/HierarchicalTrackSelectorWidget/components/faceted/util.ts
@@ -1,5 +1,0 @@
-export function getRootKeys(obj: Record<string, unknown>) {
-  return Object.entries(obj)
-    .map(([key, val]) => (typeof val === 'string' ? key : ''))
-    .filter(f => !!f)
-}
